DoctorSpecialist, Specialty c#
Repository, Service, Controller, c#
Serijalizacija + CRUD c#

+++ Zdravstena_ustanova/Model/Specialty.cs
@@ -4,10 +4,10 @@ namespace Model
 
     public class Specialty
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
 
-        public Specialty(int id, string name)
+        public Specialty(long id, string name)
         {
             Id = id;
             Name = name;
